feat: Remove useless props drilling Use redux for posts
